requirements.txt
flask
langchain
langchain-community
langchain-together
faiss-cpu
python-dotenv